+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "start:engine": "yarn --cwd packages/engine build -w",
     "start:utils": "yarn --cwd packages/utils build -w",
+    "start:server": "yarn --cwd packages/server start",
     "start:client": "yarn --cwd packages/client start",
     "start": "run-p start:*",
     "build:engine": "yarn --cwd packages/engine build",
