+++ packages/server/tsconfig.json
@@ -6,7 +6,8 @@
     "removeComments": true,
     "preserveConstEnums": true,
     "strictNullChecks": true,
-    "outDir": "dist"
+    "outDir": "dist",
+    "declaration": true
   },
   "include": [
     "**/src/**/*"
